pyproject.toml
[tool.poetry]
name = "django-notifications"
description = "Django notifications"
version = "0.0.1"
authors = ["Domenik Jones <domenik.jones@gmail.com>"]
packages = [{"include" = "notifications"}]

[tool.poetry.dependencies]
# runtime and package dependencies
poetry = ">=2.1.3"
python = ">=3.10,<3.11"

# 3rd party apps
django = ">=5.2.4"
requests = ">=2.32.4"
django-ckeditor = "^6.7.3"

[tool.poetry.group.dev.dependencies]
django-stubs = ">=5.2.2"
mypy = ">=1.17.0"
pre-commit = ">=3.8.0"
ruff = ">=0.12.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    "docs"
]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",   # PEP8 naming convetions
    "D"    # pydocstyle
]
ignore = [
    "D100",  # missing docstring in public module
    "D101",  # missing docstring in public class
    "D104",  # missing docstring in public package
    "D105",  # missing docstring in magic method
    "D106",  # missing docstring in nested class
]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]
"test_*.py" = ["D"]